extensions: purge and delete toolbar actions (#528)
* rework "purge" action

* "restore deleted" action

* fix tests

* cleanup comments

* allow inline action names

* allow inline rules without params

* simplify bulk registration
