+++ src/app/directives/node-permanent-delete.directive.ts
@@ -25,13 +25,9 @@
 
 import { Directive, HostListener, Input } from '@angular/core';
 import { MinimalNodeEntity } from 'alfresco-js-api';
-import { MatDialog } from '@angular/material';
-import { ConfirmDialogComponent } from '@alfresco/adf-content-services';
 import { Store } from '@ngrx/store';
-
-import { AppStore } from '../store/states/app.state';
+import { AppStore } from '../store/states';
 import { PurgeDeletedNodesAction } from '../store/actions';
-import { NodeInfo } from '../store/models';
 
 @Directive({
     selector: '[acaPermanentDelete]'
@@ -43,35 +39,11 @@ export class NodePermanentDeleteDirective {
     selection: MinimalNodeEntity[];
 
     constructor(
-        private store: Store<AppStore>,
-        private dialog: MatDialog
+        private store: Store<AppStore>
     ) {}
 
     @HostListener('click')
     onClick() {
-        const dialogRef = this.dialog.open(ConfirmDialogComponent, {
-            data: {
-                title: 'APP.DIALOGS.CONFIRM_PURGE.TITLE',
-                message: 'APP.DIALOGS.CONFIRM_PURGE.MESSAGE',
-                yesLabel: 'APP.DIALOGS.CONFIRM_PURGE.YES_LABEL',
-                noLabel: 'APP.DIALOGS.CONFIRM_PURGE.NO_LABEL'
-            },
-            minWidth: '250px'
-        });
-
-        dialogRef.afterClosed().subscribe(result => {
-            if (result === true) {
-                const nodesToDelete: NodeInfo[] = this.selection.map(node => {
-                    const { name } = node.entry;
-                    const id = node.entry.nodeId || node.entry.id;
-
-                    return {
-                        id,
-                        name
-                    };
-                });
-                this.store.dispatch(new PurgeDeletedNodesAction(nodesToDelete));
-            }
-        });
+        this.store.dispatch(new PurgeDeletedNodesAction(this.selection));
     }
 }
