+++ src/assets/app.extensions.json
@@ -9,8 +9,12 @@
 
     "rules": [
         {
-            "id": "app.create.canCreateFolder",
-            "type": "app.navigation.folder.canCreate"
+            "id": "app.trashcan.hasSelection",
+            "type": "core.every",
+            "parameters": [
+                { "type": "rule", "value": "app.selection.notEmpty" },
+                { "type": "rule", "value": "app.navigation.isTrashcan" }
+            ]
         },
         {
             "id": "app.toolbar.canEditFolder",
@@ -47,25 +51,6 @@
         }
     ],
 
-    "actions": [
-        {
-            "id": "app.actions.createFolder",
-            "type": "CREATE_FOLDER"
-        },
-        {
-            "id": "app.actions.editFolder",
-            "type": "EDIT_FOLDER"
-        },
-        {
-            "id": "app.actions.download",
-            "type": "DOWNLOAD_NODES"
-        },
-        {
-            "id": "app.actions.preview",
-            "type": "VIEW_FILE"
-        }
-    ],
-
     "features": {
         "create": [
             {
@@ -74,10 +59,10 @@
                 "icon": "create_new_folder",
                 "title": "ext: Create Folder",
                 "actions": {
-                    "click": "app.actions.createFolder"
+                    "click": "CREATE_FOLDER"
                 },
                 "rules": {
-                    "enabled": "app.create.canCreateFolder"
+                    "enabled": "app.navigation.folder.canCreate"
                 }
             }
         ],
@@ -149,10 +134,10 @@
                     "title": "APP.NEW_MENU.TOOLTIPS.CREATE_FOLDER",
                     "icon": "create_new_folder",
                     "actions": {
-                        "click": "app.actions.createFolder"
+                        "click": "CREATE_FOLDER"
                     },
                     "rules": {
-                        "visible": "app.create.canCreateFolder"
+                        "visible": "app.navigation.folder.canCreate"
                     }
                 },
                 {
@@ -162,7 +147,7 @@
                     "title": "APP.ACTIONS.VIEW",
                     "icon": "open_in_browser",
                     "actions": {
-                        "click": "app.actions.preview"
+                        "click": "VIEW_FILE"
                     },
                     "rules": {
                         "visible": "app.toolbar.canViewFile"
@@ -175,7 +160,7 @@
                     "title": "APP.ACTIONS.DOWNLOAD",
                     "icon": "get_app",
                     "actions": {
-                        "click": "app.actions.download"
+                        "click": "DOWNLOAD_NODES"
                     },
                     "rules": {
                         "visible": "app.toolbar.canDownload"
@@ -188,10 +173,34 @@
                     "title": "APP.ACTIONS.EDIT",
                     "icon": "create",
                     "actions": {
-                        "click": "app.actions.editFolder"
+                        "click": "EDIT_FOLDER"
                     },
                     "rules": {
                         "visible": "app.toolbar.canEditFolder"
+                    }
+                },
+                {
+                    "id": "app.toolbar.purgeDeletedNodes",
+                    "type": "button",
+                    "title": "APP.ACTIONS.DELETE_PERMANENT",
+                    "icon": "delete_forever",
+                    "actions": {
+                        "click": "PURGE_DELETED_NODES"
+                    },
+                    "rules": {
+                        "visible": "app.trashcan.hasSelection"
+                    }
+                },
+                {
+                    "id": "app.toolbar.restoreDeletedNodes",
+                    "type": "button",
+                    "title": "APP.ACTIONS.RESTORE",
+                    "icon": "restore",
+                    "actions": {
+                        "click": "RESTORE_DELETED_NODES"
+                    },
+                    "rules": {
+                        "visible": "app.trashcan.hasSelection"
                     }
                 },
                 {
